WIP
I was trying to use `:focus:not(:focus-visible)` to only apply the focus ring when using the keyboard. Unfortunately, it isn’t working as expected.

I was diggin in to how the CSS gets applied to the page. Perhaps the way EmotionCSS and Vite CSS imports work are messing with the way the CSS is applied.